[buck2]REFACTOR: Integrate orion and orion server with generic api and model interface
