+++ Cargo.toml
@@ -140,6 +140,7 @@ redis-test = "1.0.1"
 bincode = "2.0.1"
 rustls = "0.23.36"
 object_store = "0.13.1"
+parse_display = "0.8.2"
 qlean = "0.1.1"
 
 [profile.release]
