+++ api-model/src/buck2/mod.rs
@@ -1,3 +1,4 @@
 pub mod api;
 pub mod types;
 pub mod ws;
+pub mod status;
