+++ jupiter/callisto/src/build_events.rs
@@ -1,4 +1,4 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.19
+/// User-defined build event, should be changed later in migration to fit
 
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
@@ -8,11 +8,13 @@ use serde::{Deserialize, Serialize};
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
     pub id: Uuid,
-    pub index: i32,
     pub task_id: Uuid,
     pub exit_code: Option<i32>,
+    pub retry_count: i32,
+    pub repo: String,
+    pub log: Option<String>,
     pub log_output_file: String,
-    pub start_at: Option<DateTimeWithTimeZone>,
+    pub start_at: DateTimeWithTimeZone,
     pub end_at: Option<DateTimeWithTimeZone>,
 }
 
@@ -35,3 +37,21 @@ impl Related<super::tasks::Entity> for Entity {
 }
 
 impl ActiveModelBehavior for ActiveModel {}
+
+trait BuildModel {
+    /// Crete a new build event ActiveModel for database insertion
+    fn create_build_event(
+        build_event_id: Uuid,
+        task_id: Uuid,
+        repo: String,
+        retry_count: i32,
+    ) -> ActiveModel;
+
+    fn insert_build_event(
+        build_event_id: Uuid,
+        task_id: Uuid,
+        repo: String,
+        retry_count: i32,
+        db: &impl ConnectionTrait,
+    ) -> Result<Model, DbErr>;
+}
