+++ orion-server/Cargo.toml
@@ -8,6 +8,7 @@ orion = { workspace = true }
 common = { workspace = true}
 io-orbit = { workspace = true }
 api-model = { workspace = true }
+jupiter = {workspace = true}
 
 http = { workspace = true }
 axum = { workspace = true, features = ["macros", "ws"] }
