+++ orion/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+api-model = { workspace = true }
 tokio = { workspace = true, features = ["rt-multi-thread", "fs", "process"] }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
