+++ orion/src/repo/changes.rs
@@ -10,13 +10,13 @@
 
 use std::collections::HashSet;
 
+use api_model::buck2::{status::Status, types::ProjectRelativePath};
 use td_util::prelude::*;
 use td_util_buck::{
     cells::CellInfo,
-    types::{CellPath, Package, ProjectRelativePath},
+    types::{CellPath, Package},
 };
-
-use crate::repo::sapling::status::Status;
+use ProjectRelativePath;
 
 #[derive(Default, Debug)]
 pub struct Changes {
@@ -93,9 +93,7 @@ impl Changes {
 #[cfg(test)]
 mod tests {
     use td_util_buck::types::{CellPath, ProjectRelativePath};
-
     use super::*;
-    use crate::repo::sapling::status::Status;
 
     #[test]
     fn test_changes_empty() {
